Increased the wait for couchbase initialisation.
CouchBase takes a moment to complete the initialisation after receiving the
command via its API.  The duration of this pause is hardware-dependent and it
was not long enough for everyone's equipment.

This increases the default and exposes it as an env var.  Given this pause only
comes about during the initialisation of a new couchbase install, perhaps the
env var is overkill...